move standard functionality to its own module
